requirements.txt
openai
pandas
langchain
streamlit
python-dotenv
openpyxl